+++ src/components/index.js
@@ -71,6 +71,7 @@ export { default as Snackbar } from './snackbar/Snackbar';
 export { default as Skeleton } from './skeleton';
 export { snackbar } from './snackbar';
 export { default as Sidebar } from './sidebar';
+export { default as Spinner } from './spinner';
 export { ThemeContext } from './theme-provider';
 export { 
     convertToNum,