+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog 
 
+#### `0.9.1` `2021-01-14`
+- `NEW!` Spinner component added;
+- Button styles updated: set icon custom color on light mode;
+
 #### `0.9.0` `2021-01-13`
 - Components demo migrated to Storybook
 - Types wrong declarations fixed (TS, prop-types);